+++ package.json
@@ -22,6 +22,7 @@
     "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
     "eslint-plugin-vue": "^6.2.2",
+    "vue-router": "^3.5.1",
     "vue-template-compiler": "^2.6.11"
   },
   "eslintConfig": {
