+++ src/main.js
@@ -1,8 +1,11 @@
 import Vue from 'vue'
 import App from './App.vue'
+// import "./firebase.js"
+import router from './router'
 
 Vue.config.productionTip = false
 
 new Vue({
-  render: h => h(App),
+  router,
+  render: h => h(App)
 }).$mount('#app')
