+++ src/pages/Hoge.vue
@@ -1,0 +1,10 @@
+<template>
+  <div>This is Hoge</div>
+</template>
+
+<script>
+export default {
+  
+}
+</script>
+
